+++ src/EnrolmentWorkflow/Action/AbstractAction.php
@@ -209,7 +209,7 @@ abstract class AbstractAction extends ActiveRecord
                 return ($field_value ? 1 : 0);
 
             default:
-                return null;
+                return parent::sleep($field_name);
         }
     }
 
@@ -225,7 +225,7 @@ abstract class AbstractAction extends ActiveRecord
                 return boolval($field_value);
 
             default:
-                return null;
+                return parent::wakeUp($field_name, $field_value);
         }
     }