+++ src/EnrolmentWorkflow/Member/Member.php
@@ -184,7 +184,7 @@ class Member extends ActiveRecord
                 return json_encode($field_value);
 
             default:
-                return null;
+                return parent::sleep($field_name);
         }
     }
 
@@ -199,7 +199,7 @@ class Member extends ActiveRecord
                 return json_decode($field_value, true);
 
             default:
-                return null;
+                return parent::wakeUp($field_name, $field_value);
         }
     }