+++ src/EnrolmentWorkflow/Request/Request.php
@@ -305,7 +305,7 @@ class Request extends ActiveRecord
                 return json_encode($field_value);
 
             default:
-                return null;
+                return parent::sleep($field_name);
         }
     }
 
@@ -323,7 +323,7 @@ class Request extends ActiveRecord
                 return json_decode($field_value);
 
             default:
-                return null;
+                return parent::wakeUp($field_name, $field_value);
         }
     }