+++ src/EnrolmentWorkflow/Rule/AbstractRule.php
@@ -231,7 +231,7 @@ abstract class AbstractRule extends ActiveRecord
                 return ($field_value ? 1 : 0);
 
             default:
-                return null;
+                return parent::sleep($field_name);
         }
     }
 
@@ -246,7 +246,7 @@ abstract class AbstractRule extends ActiveRecord
                 return boolval($field_value);
 
             default:
-                return null;
+                return parent::wakeUp($field_name, $field_value);
         }
     }