+++ src/EnrolmentWorkflow/Step/Step.php
@@ -260,7 +260,7 @@ class Step extends ActiveRecord
                 return json_encode($field_value);
 
             default:
-                return null;
+                return parent::sleep($field_name);
         }
     }
 
@@ -280,7 +280,7 @@ class Step extends ActiveRecord
                 return json_decode($field_value, true);
 
             default:
-                return null;
+                return parent::wakeUp($field_name, $field_value);
         }
     }