+++ src/RuleEnrolment/Rule/Rule.php
@@ -210,7 +210,7 @@ class Rule extends ActiveRecord
 
         switch ($field_name) {
             default:
-                return null;
+                return parent::sleep($field_name);
         }
     }
 
@@ -237,7 +237,7 @@ class Rule extends ActiveRecord
                 return boolval($field_value);
 
             default:
-                return null;
+                return parent::wakeUp($field_name, $field_value);
         }
     }